Refactors event service and updates navigation structure
Renames mockTickets to createRandomEvent for clarity and updates the routing structure to include a new QR scanner component under the events route.

Enhances the QrsComponent to implement modal functionality for event details and modifies the NavBarMenuComponent to use signals for better state management.

Improves user navigation and interaction within the application.